Changed the literal translation of "blade"
I don't know if "Blatt" is the right translation for "blade".  It sounds strange and I would not use it.  So in this example, I'm changing the translation to avoid the term "blade" altogether, as I don't think it's important to refer to the blade itself when describing the action.  Again, I think there might be a better, more appropriate translation for "blade".